+++ packages/analytics/analytics-chart/src/enums/chart-legend-position.enum.ts
@@ -1,6 +1,5 @@
 
 export enum ChartLegendPosition {
   Hidden = 'hidden',
-  Right = 'right',
   Bottom = 'bottom',
 }