+++ packages/entities/entities-plugins/src/components/free-form/Datakit/examples/index.ts
@@ -1,3 +1,13 @@
-export { default as authenticate } from './authenticate.yml?raw'
-export { default as combine } from './combine.yml?raw'
-export { default as manipulate } from './manipulate.yml?raw'
+import { default as authenticate } from './authenticate.yml?raw'
+import { default as vault } from './vault.yml?raw'
+import { default as combine } from './combine.yml?raw'
+import { default as cache } from './cache.yml?raw'
+import { default as manipulate } from './manipulate.yml?raw'
+
+export default {
+  authenticate,
+  vault,
+  combine,
+  cache,
+  manipulate,
+}