+++ packages/entities/entities-plugins/src/components/free-form/shared/layout/StandardLayout.vue
@@ -73,7 +73,7 @@
           />
         </div>
         <div
-          v-if="scoped"
+          v-if="scoped && scopeEntitiesSchema"
           class="scope-detail"
         >
           <VFGField
@@ -235,29 +235,64 @@ const realFormConfig = computed(() => {
   }
 })
 
+interface LegacyFormSchemaField {
+  model: string
+  type: string
+  fields: LegacyFormSchemaField[]
+  label?: string
+  description?: string
+}
+
+interface LegacyFormSchemaGroup {
+  fields: LegacyFormSchemaField[]
+  collapsible?: {
+    nestedCollapsible?: {
+      fields?: LegacyFormSchemaField[]
+    }
+  }
+}
+
+const flattenFormSchemaFields = computed<LegacyFormSchemaField[]>(() => {
+  if (props.formSchema?.groups) {
+    return props.formSchema.groups.reduce(
+      (acc: any[], group: LegacyFormSchemaGroup) => {
+        const fields = group.fields
+
+        const allFields = group.collapsible?.nestedCollapsible?.fields
+          ? fields.concat(group.collapsible.nestedCollapsible.fields)
+          : fields
+
+        return acc.concat(allFields)
+      },
+      [] as LegacyFormSchemaField[],
+    )
+  }
+  return props.formSchema?.fields || []
+})
+
 const enabledSchema = computed(() => {
   return {
-    fields: props.formSchema?.fields.filter((field: { model: string }) => field.model === 'enabled'),
+    fields: flattenFormSchemaFields.value.filter(field => field.model === 'enabled'),
   }
 })
 
 const scopeSchema = computed(() => {
-  return props.formSchema?.fields.find((field: { model: string }) => field.model === 'selectionGroup')
+  return flattenFormSchemaFields.value.find(field => field.type === 'selectionGroup')
 })
 
 const radioGroup = computed(() => {
-  return (scopeSchema.value.fields || []).map(({ label, description }: { label: string, description: string }) => ({ label, description }))
+  return (scopeSchema.value?.fields || []).map(({ label, description }) => ({ label, description }))
 })
 
 const scopeEntitiesSchema = computed(() => {
-  return scopeSchema.value.fields[1]
+  return scopeSchema.value?.fields?.[1]
 })
 
 const moreFieldsSchema = computed(() => {
   const fields = ['instance_name', 'protocols', 'tags']
 
   return {
-    fields: props.formSchema?.fields.filter((field: { model: string }) => fields.includes(field.model)),
+    fields: flattenFormSchemaFields.value.filter(field => fields.includes(field.model!)),
   }
 })
 
@@ -273,7 +308,7 @@ const freeFormSchema = computed(() => {
     // enabled field
     ...enabledSchema.value.fields,
     // scope fields
-    ...scopeEntitiesSchema.value.fields,
+    ...(scopeEntitiesSchema.value?.fields ?? []),
     // general info fields
     ...moreFieldsSchema.value.fields,
   ]
@@ -356,7 +391,7 @@ const prunedData = computed(() => {
 
 
 const scopeFields = computed<Array<'service' | 'route' | 'consumer' | 'consumer_group'>>(() => {
-  return scopeEntitiesSchema.value.fields.map((field: any) => field.model.split('-')[0])
+  return scopeEntitiesSchema.value?.fields.map((field: any) => field.model.split('-')[0]) ?? []
 })
 
 // Cache of scope-related fields to restore when toggling scoped on/off